chore#451: 도메인 재배치
