+++ src/main/java/or/sopt/houme/domain/banner/presentation/dto/response/BannerExploreResponse.java
@@ -1,4 +1,4 @@
-package or.sopt.houme.domain.explore.presentation.dto.response;
+package or.sopt.houme.domain.banner.presentation.dto.response;
 
 import or.sopt.houme.domain.banner.model.entity.Banner;
