+++ src/main/java/or/sopt/houme/domain/banner/presentation/dto/response/LandingListResponse.java
@@ -1,0 +1,12 @@
+package or.sopt.houme.domain.banner.presentation.dto.response;
+
+import java.util.List;
+
+public record LandingListResponse(
+        List<LandingResponse> landings
+) {
+
+    public static LandingListResponse of(List<LandingResponse> landings) {
+        return new LandingListResponse(landings);
+    }
+}
