--- src/main/java/or/sopt/houme/domain/explore/presentation/controller/ExploreController.java
@@ -1,97 +1,0 @@
-package or.sopt.houme.domain.explore.presentation.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import or.sopt.houme.domain.explore.presentation.dto.response.BannerDetailResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.BannerExploreListResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.ExploreHouseTemplateDetailResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.ExploreHouseTemplateListResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.OtherStyleDetailResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.OtherStyleListResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.RecentFloorPlanResponse;
-import or.sopt.houme.domain.explore.service.ExploreService;
-import or.sopt.houme.domain.house.model.entity.enums.Equilibrium;
-import or.sopt.houme.domain.house.model.entity.enums.Form;
-import or.sopt.houme.domain.house.model.entity.enums.Structure;
-import or.sopt.houme.domain.user.presentation.controller.dto.CustomUserDetails;
-import or.sopt.houme.global.api.ApiResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/explore")
-@RequiredArgsConstructor
-@Tag(name = "탐색 탭 관련 API")
-public class ExploreController {
-
-    private final ExploreService exploreService;
-
-    @GetMapping("/banners/{bannerId}")
-    @Operation(summary = "배너 전체 조회 API",
-            description = "요청한 bannerId를 첫 번째로 두고 circular 방식으로 배너 목록을 반환합니다")
-    public ResponseEntity<ApiResponse<BannerExploreListResponse>> getExploreBanners(@PathVariable Long bannerId) {
-        return ResponseEntity.ok(ApiResponse.ok(exploreService.getExploreBanners(bannerId)));
-    }
-
-    @GetMapping("/banners/{bannerId}/detail")
-    @Operation(summary = "배너 디테일 페이지 조회 API")
-    public ResponseEntity<ApiResponse<BannerDetailResponse>> getExploreBannerDetail(@PathVariable Long bannerId) {
-        return ResponseEntity.ok(ApiResponse.ok(exploreService.getExploreBannerDetail(bannerId)));
-    }
-
-    @GetMapping("/other-styles")
-    @Operation(summary = "다른 스타일 전체 조회 API")
-    public ResponseEntity<ApiResponse<OtherStyleListResponse>> getOtherStyles(
-            @RequestParam(required = false) Integer size
-    ) {
-        return ResponseEntity.ok(ApiResponse.ok(exploreService.getOtherStyles(size)));
-    }
-
-    @GetMapping("/other-styles/{styleId}")
-    @Operation(summary = "스타일 디테일 페이지 조회 API")
-    public ResponseEntity<ApiResponse<OtherStyleDetailResponse>> getOtherStyleDetail(@PathVariable Long styleId) {
-        return ResponseEntity.ok(ApiResponse.ok(exploreService.getOtherStyleDetail(styleId)));
-    }
-
-    @GetMapping("/house-templates")
-    @Operation(summary = "도면 전체 조회 API")
-    public ResponseEntity<ApiResponse<ExploreHouseTemplateListResponse>> getExploreHouseTemplates(
-            @RequestParam(required = false) Integer size,
-            @RequestParam(required = false) Form residenceType,
-            @RequestParam(required = false) Structure layoutType,
-            @RequestParam(required = false) Equilibrium equilibrium,
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        return ResponseEntity.ok(ApiResponse.ok(
-                exploreService.getExploreHouseTemplates(
-                        size,
-                        residenceType,
-                        layoutType,
-                        equilibrium,
-                        userDetails != null ? userDetails.getUser() : null
-                )
-        ));
-    }
-
-    @GetMapping("/house-templates/{floorPlanId}")
-    @Operation(summary = "도면 상세 조회 API")
-    public ResponseEntity<ApiResponse<ExploreHouseTemplateDetailResponse>> getExploreHouseTemplateDetail(
-            @PathVariable Long floorPlanId
-    ) {
-        return ResponseEntity.ok(ApiResponse.ok(exploreService.getExploreHouseTemplateDetail(floorPlanId)));
-    }
-
-    @GetMapping("/recent-floor-plan")
-    @Operation(summary = "최근 사용한 도면 데이터 조회 API")
-    public ResponseEntity<ApiResponse<RecentFloorPlanResponse>> getRecentFloorPlan(
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        return ResponseEntity.ok(ApiResponse.ok(exploreService.getRecentFloorPlan(userDetails.getUser())));
-    }
-}
