--- src/main/java/or/sopt/houme/domain/explore/service/ExploreServiceImpl.java
@@ -1,358 +1,0 @@
-package or.sopt.houme.domain.explore.service;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import or.sopt.houme.domain.banner.model.entity.Banner;
-import or.sopt.houme.domain.banner.model.entity.BannerCurationRawProduct;
-import or.sopt.houme.domain.banner.model.entity.BannerType;
-import or.sopt.houme.domain.banner.model.vo.BannerStyleAnswerChip;
-import or.sopt.houme.domain.banner.repository.BannerRepository;
-import or.sopt.houme.domain.explore.presentation.dto.response.ExploreHouseTemplateDetailItemResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.ExploreHouseTemplateDetailResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.RecentFloorPlanItemResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.RecentFloorPlanResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.BannerDetailAnswerResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.BannerDetailResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.BannerExploreListResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.BannerExploreResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.ExploreHouseTemplateItemResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.ExploreHouseTemplateListResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.OtherStyleDetailProductResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.OtherStyleDetailResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.OtherStyleListResponse;
-import or.sopt.houme.domain.explore.presentation.dto.response.OtherStyleResponse;
-import or.sopt.houme.domain.generateImage.model.entity.GenerateImage;
-import or.sopt.houme.domain.generateImage.repository.GenerateImageRepository;
-import or.sopt.houme.domain.house.model.entity.enums.Equilibrium;
-import or.sopt.houme.domain.house.model.entity.enums.Form;
-import or.sopt.houme.domain.house.model.entity.enums.Structure;
-import or.sopt.houme.domain.house.model.entity.mapping.HouseFloorPlan;
-import or.sopt.houme.domain.house.model.floorPlan.entity.FloorPlan;
-import or.sopt.houme.domain.house.model.floorPlan.vo.FloorPlanImageItem;
-import or.sopt.houme.domain.house.repository.floorPlan.FloorPlanRepository;
-import or.sopt.houme.domain.user.model.entity.User;
-import or.sopt.houme.domain.user.util.floorplan.FloorPlanImageJsonCodec;
-import or.sopt.houme.global.api.ErrorCode;
-import or.sopt.houme.global.api.GeneralException;
-import or.sopt.houme.global.api.handler.BannerException;
-import or.sopt.houme.global.api.handler.HouseException;
-import or.sopt.houme.global.api.handler.ValidException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class ExploreServiceImpl implements ExploreService {
-
-    private static final TypeReference<List<BannerStyleAnswerChip>> STYLE_ANSWER_CHIP_TYPE = new TypeReference<>() {};
-
-    private final BannerRepository bannerRepository;
-    private final GenerateImageRepository generateImageRepository;
-    private final FloorPlanRepository floorPlanRepository;
-    private final FloorPlanImageJsonCodec floorPlanImageJsonCodec;
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public BannerExploreListResponse getExploreBanners(Long bannerId) {
-        List<Banner> banners = bannerRepository.findAllWithRawProducts(BannerType.BANNER, false).stream()
-                .sorted((left, right) -> Long.compare(left.getId(), right.getId()))
-                .toList();
-        int startIndex = findBannerStartIndex(banners, bannerId);
-
-        List<BannerExploreResponse> orderedBanners = new ArrayList<>(banners.size());
-        for (int index = 0; index < banners.size(); index++) {
-            Banner banner = banners.get((startIndex + index) % banners.size());
-            orderedBanners.add(BannerExploreResponse.from(banner));
-        }
-        return BannerExploreListResponse.of(orderedBanners);
-    }
-
-    @Override
-    public BannerDetailResponse getExploreBannerDetail(Long bannerId) {
-        Banner banner = bannerRepository.findByIdWithRawProducts(bannerId, BannerType.BANNER, false)
-                .orElseThrow(() -> new BannerException(ErrorCode.NOT_FOUND_BANNER));
-
-        return BannerDetailResponse.of(
-                banner.getBannerTitle(),
-                banner.getBannerImageUrl(),
-                banner.getStyleQuestion(),
-                parseStyleAnswerChips(banner.getStyleAnswerChipsJson()).stream()
-                        .map(chip -> BannerDetailAnswerResponse.of(chip.label()))
-                        .toList()
-        );
-    }
-
-    @Override
-    public OtherStyleListResponse getOtherStyles(Integer size) {
-        if (size != null && size < 1) {
-            throw new ValidException(ErrorCode.NOT_VALID_EXCEPTION);
-        }
-
-        List<OtherStyleResponse> styles = bannerRepository.findAllWithRawProducts(BannerType.STYLE, false).stream()
-                .sorted((left, right) -> Long.compare(left.getId(), right.getId()))
-                .map(OtherStyleResponse::from)
-                .toList();
-
-        if (size == null) {
-            return OtherStyleListResponse.of(styles);
-        }
-        return OtherStyleListResponse.of(styles.stream().limit(size).toList());
-    }
-
-    @Override
-    public OtherStyleDetailResponse getOtherStyleDetail(Long styleId) {
-        Banner style = bannerRepository.findByIdWithRawProducts(styleId, BannerType.STYLE, false)
-                .orElseThrow(() -> new BannerException(ErrorCode.NOT_FOUND_STYLE));
-
-        List<OtherStyleDetailProductResponse> products = style.getBannerRawProducts().stream()
-                .sorted((left, right) -> Long.compare(safeMappingId(left), safeMappingId(right)))
-                .map(BannerCurationRawProduct::getCurationRawProduct)
-                .filter(java.util.Objects::nonNull)
-                .map(OtherStyleDetailProductResponse::from)
-                .toList();
-
-        return OtherStyleDetailResponse.of(
-                style.getBannerTitle(),
-                style.getBannerImageUrl(),
-                style.getStyleDescription(),
-                products
-        );
-    }
-
-    @Override
-    public RecentFloorPlanResponse getRecentFloorPlan(User user) {
-        Optional<GenerateImage> recentGenerateImage = generateImageRepository.findMostRecentByUserId(user.getId());
-        if (recentGenerateImage.isEmpty()) {
-            return RecentFloorPlanResponse.noRecent();
-        }
-
-        FloorPlan floorPlan = recentGenerateImage.get().getHouse().getHouseFloorPlans().stream()
-                .sorted((left, right) -> Long.compare(safeHouseFloorPlanId(left), safeHouseFloorPlanId(right)))
-                .map(HouseFloorPlan::getFloorPlan)
-                .filter(java.util.Objects::nonNull)
-                .findFirst()
-                .orElse(null);
-
-        if (floorPlan == null) {
-            return RecentFloorPlanResponse.noRecent();
-        }
-
-        FloorPlanImageItem representativeImage = resolveRepresentativeFloorPlanImage(floorPlan);
-        return RecentFloorPlanResponse.withRecent(
-                RecentFloorPlanItemResponse.of(floorPlan, representativeImage.url(), representativeImage.view())
-        );
-    }
-
-    @Override
-    public ExploreHouseTemplateListResponse getExploreHouseTemplates(
-            Integer size,
-            Form residenceType,
-            Structure layoutType,
-            Equilibrium equilibrium,
-            User user
-    ) {
-        validateSize(size);
-
-        List<FloorPlan> allFloorPlans = floorPlanRepository.findAll().stream()
-                .sorted((left, right) -> Long.compare(left.getId(), right.getId()))
-                .toList();
-
-        List<FloorPlan> exactFloorPlans = allFloorPlans.stream()
-                .filter(floorPlan -> matchesExactFilter(floorPlan, residenceType, layoutType, equilibrium))
-                .toList();
-
-        boolean isExact = !exactFloorPlans.isEmpty();
-        List<FloorPlan> selectedFloorPlans = isExact
-                ? exactFloorPlans
-                : recommendSimilarFloorPlans(allFloorPlans, residenceType, layoutType);
-
-        List<FloorPlan> limitedFloorPlans = applySize(selectedFloorPlans, size);
-        Long latestFloorPlanId = resolveLatestFloorPlanId(user);
-
-        List<ExploreHouseTemplateItemResponse> responses = limitedFloorPlans.stream()
-                .map(floorPlan -> {
-                    FloorPlanImageItem representative = resolveRepresentativeFloorPlanImage(floorPlan);
-                    return ExploreHouseTemplateItemResponse.of(
-                            floorPlan,
-                            representative.url(),
-                            latestFloorPlanId != null && latestFloorPlanId.equals(floorPlan.getId())
-                    );
-                })
-                .toList();
-
-        return ExploreHouseTemplateListResponse.of(isExact, responses);
-    }
-
-    @Override
-    public ExploreHouseTemplateDetailResponse getExploreHouseTemplateDetail(Long floorPlanId) {
-        FloorPlan floorPlan = floorPlanRepository.findById(floorPlanId)
-                .orElseThrow(() -> new HouseException(ErrorCode.NOT_FOUND_FLOOR_PLAN));
-
-        List<FloorPlanImageItem> images = floorPlanImageJsonCodec.read(floorPlan.getImagesJson());
-        if (images.isEmpty()) {
-            images = List.of(FloorPlanImageItem.create(
-                    floorPlan.getUrl(),
-                    floorPlan.getFilename(),
-                    floorPlan.getOriginalFilename(),
-                    floorPlan.getFileExtension(),
-                    1,
-                    null
-            ));
-        }
-
-        List<ExploreHouseTemplateDetailItemResponse> responses = images.stream()
-                .filter(image -> image.url() != null && !image.url().isBlank())
-                .map(image -> ExploreHouseTemplateDetailItemResponse.of(image.url(), image.view()))
-                .toList();
-
-        return ExploreHouseTemplateDetailResponse.of(
-                floorPlan.getId(),
-                floorPlan.getFloorPlanName(),
-                floorPlan.getEquilibrium() != null ? floorPlan.getEquilibrium().getDescription() : null,
-                responses
-        );
-    }
-
-    private int findBannerStartIndex(List<Banner> banners, Long bannerId) {
-        for (int index = 0; index < banners.size(); index++) {
-            if (banners.get(index).getId().equals(bannerId)) {
-                return index;
-            }
-        }
-        throw new BannerException(ErrorCode.NOT_FOUND_BANNER);
-    }
-
-    private List<BannerStyleAnswerChip> parseStyleAnswerChips(String styleAnswerChipsJson) {
-        if (styleAnswerChipsJson == null || styleAnswerChipsJson.isBlank()) {
-            return List.of();
-        }
-        try {
-            List<BannerStyleAnswerChip> chips = objectMapper.readValue(styleAnswerChipsJson, STYLE_ANSWER_CHIP_TYPE);
-            if (chips == null) {
-                return List.of();
-            }
-            return chips.stream()
-                    .sorted((left, right) -> Integer.compare(left.order(), right.order()))
-                    .toList();
-        } catch (JsonProcessingException e) {
-            throw new GeneralException(ErrorCode.OBJECTMAPPER_EXCEPTION);
-        }
-    }
-
-    private long safeMappingId(BannerCurationRawProduct mapping) {
-        if (mapping == null || mapping.getId() == null) {
-            return Long.MAX_VALUE;
-        }
-        return mapping.getId();
-    }
-
-    private long safeHouseFloorPlanId(HouseFloorPlan mapping) {
-        if (mapping == null || mapping.getId() == null) {
-            return Long.MAX_VALUE;
-        }
-        return mapping.getId();
-    }
-
-    private FloorPlanImageItem resolveRepresentativeFloorPlanImage(FloorPlan floorPlan) {
-        return floorPlanImageJsonCodec.read(floorPlan.getImagesJson()).stream()
-                .filter(item -> item.url() != null && !item.url().isBlank())
-                .findFirst()
-                .orElseGet(() -> FloorPlanImageItem.create(
-                        floorPlan.getUrl(),
-                        floorPlan.getFilename(),
-                        floorPlan.getOriginalFilename(),
-                        floorPlan.getFileExtension(),
-                        1,
-                        null
-                ));
-    }
-
-    private void validateSize(Integer size) {
-        if (size != null && size < 1) {
-            throw new ValidException(ErrorCode.NOT_VALID_EXCEPTION);
-        }
-    }
-
-    private boolean matchesExactFilter(
-            FloorPlan floorPlan,
-            Form residenceType,
-            Structure layoutType,
-            Equilibrium equilibrium
-    ) {
-        if (residenceType != null && floorPlan.getForm() != residenceType) {
-            return false;
-        }
-        if (layoutType != null && floorPlan.getStructure() != layoutType) {
-            return false;
-        }
-        if (equilibrium != null && floorPlan.getEquilibrium() != equilibrium) {
-            return false;
-        }
-        return true;
-    }
-
-    private List<FloorPlan> recommendSimilarFloorPlans(
-            List<FloorPlan> allFloorPlans,
-            Form residenceType,
-            Structure layoutType
-    ) {
-        if (residenceType == null && layoutType == null) {
-            return allFloorPlans;
-        }
-
-        List<FloorPlan> similar = allFloorPlans.stream()
-                .filter(floorPlan ->
-                        (residenceType != null && floorPlan.getForm() == residenceType)
-                                || (layoutType != null && floorPlan.getStructure() == layoutType))
-                .toList();
-
-        if (!similar.isEmpty()) {
-            return deduplicateById(similar);
-        }
-        return allFloorPlans;
-    }
-
-    private List<FloorPlan> applySize(List<FloorPlan> floorPlans, Integer size) {
-        if (size == null) {
-            return floorPlans;
-        }
-        return floorPlans.stream().limit(size).toList();
-    }
-
-    private Long resolveLatestFloorPlanId(User user) {
-        if (user == null) {
-            return null;
-        }
-
-        return generateImageRepository.findMostRecentByUserId(user.getId())
-                .map(GenerateImage::getHouse)
-                .flatMap(house -> house.getHouseFloorPlans().stream()
-                        .sorted((left, right) -> Long.compare(safeHouseFloorPlanId(left), safeHouseFloorPlanId(right)))
-                        .map(HouseFloorPlan::getFloorPlan)
-                        .filter(java.util.Objects::nonNull)
-                        .map(FloorPlan::getId)
-                        .findFirst())
-                .orElse(null);
-    }
-
-    private List<FloorPlan> deduplicateById(List<FloorPlan> floorPlans) {
-        Map<Long, FloorPlan> byId = floorPlans.stream()
-                .collect(Collectors.toMap(
-                        FloorPlan::getId,
-                        floorPlan -> floorPlan,
-                        (left, right) -> left,
-                        LinkedHashMap::new
-                ));
-        return List.copyOf(byId.values());
-    }
-}
