+++ src/main/java/or/sopt/houme/domain/house/presentation/floorPlan/dto/response/ExploreHouseTemplateDetailResponse.java
@@ -1,4 +1,4 @@
-package or.sopt.houme.domain.explore.presentation.dto.response;
+package or.sopt.houme.domain.house.presentation.floorPlan.dto.response;
 
 import java.util.List;
