+++ src/main/java/or/sopt/houme/domain/house/presentation/floorPlan/dto/response/ExploreHouseTemplateItemResponse.java
@@ -1,4 +1,4 @@
-package or.sopt.houme.domain.explore.presentation.dto.response;
+package or.sopt.houme.domain.house.presentation.floorPlan.dto.response;
 
 import or.sopt.houme.domain.house.model.floorPlan.entity.FloorPlan;
