+++ src/main/java/or/sopt/houme/domain/user/service/UserLandingServiceImpl.java
@@ -3,11 +3,7 @@ package or.sopt.houme.domain.user.service;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import or.sopt.houme.domain.banner.model.entity.BannerType;
-import or.sopt.houme.domain.banner.repository.BannerRepository;
 import or.sopt.houme.domain.user.model.entity.User;
-import or.sopt.houme.domain.user.presentation.controller.dto.LandingListResponse;
-import or.sopt.houme.domain.user.presentation.controller.dto.LandingResponse;
 import or.sopt.houme.domain.user.repository.RefreshTokenRepository;
 import or.sopt.houme.domain.user.repository.UserRepository;
 import or.sopt.houme.global.api.ErrorCode;
@@ -17,8 +13,6 @@ import or.sopt.houme.global.jwt.JWTUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -27,17 +21,6 @@ public class UserLandingServiceImpl implements UserLandingService {
     private final UserRepository userRepository;
     private final JWTUtil jwtUtil;
     private final RefreshTokenRepository refreshTokenRepository;
-    private final BannerRepository bannerRepository;
-
-    @Override
-    public LandingListResponse getLandings() {
-        return LandingListResponse.of(
-                bannerRepository.findAllWithRawProducts(BannerType.BANNER, false).stream()
-                        .sorted((left, right) -> Long.compare(left.getId(), right.getId()))
-                        .map(LandingResponse::from)
-                        .toList()
-        );
-    }
 
     @Override
     public Boolean getHasGeneratedImage(HttpServletRequest request){
