+++ src/test/java/or/sopt/houme/domain/user/service/UserLandingServiceImplTest.java
@@ -2,11 +2,7 @@ package or.sopt.houme.domain.user.service;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
-import or.sopt.houme.domain.banner.model.entity.Banner;
-import or.sopt.houme.domain.banner.model.entity.BannerType;
-import or.sopt.houme.domain.banner.repository.BannerRepository;
 import or.sopt.houme.domain.user.model.entity.User;
-import or.sopt.houme.domain.user.presentation.controller.dto.LandingListResponse;
 import or.sopt.houme.domain.user.repository.RefreshTokenRepository;
 import or.sopt.houme.domain.user.repository.UserRepository;
 import or.sopt.houme.global.api.ErrorCode;
@@ -20,7 +16,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -38,35 +33,8 @@ class UserLandingServiceImplTest {
     @Mock
     private RefreshTokenRepository refreshTokenRepository;
 
-    @Mock
-    private BannerRepository bannerRepository;
-
     @InjectMocks
     private UserLandingServiceImpl userLandingService;
-
-    @Test
-    @DisplayName("랜딩 목록 조회 시 배너를 랜딩 응답으로 변환한다")
-    void getLandings_returnsBanners() {
-        Banner firstBanner = Banner.builder()
-                .id(1L)
-                .bannerTitle("재택근무가 필요한")
-                .bannerImageUrl("https://google.com")
-                .build();
-        Banner secondBanner = Banner.builder()
-                .id(2L)
-                .bannerTitle("취향 탐색이 필요한")
-                .bannerImageUrl("https://google.com")
-                .build();
-        when(bannerRepository.findAllWithRawProducts(BannerType.BANNER, false))
-                .thenReturn(List.of(firstBanner, secondBanner));
-
-        LandingListResponse result = userLandingService.getLandings();
-
-        assertEquals(2, result.landings().size());
-        assertEquals(1L, result.landings().get(0).id());
-        assertEquals("재택근무가 필요한", result.landings().get(0).name());
-        assertEquals("https://google.com", result.landings().get(0).imageUrl());
-    }
 
     @Test
     @DisplayName("쿠키가 없으면 true 반환")
